clean up and fix circular imports amend: fix missing import
