+++ bayesflow/experimental/networks/coupling_flow/transforms/affine_transform.py
