--- bayesflow/experimental/networks/helper_networks/__init__.py
@@ -1,2 +1,0 @@
-
-from .residual import ConditionalResidualBlock
