--- bayesflow/experimental/networks/inference_networks/__init__.py
@@ -1,3 +1,0 @@
-
-from .coupling_flow import CouplingFlow
-# from .flow_matching import FlowMatching
