--- bayesflow/experimental/networks/summary_networks/summary_model.py
@@ -1,6 +1,0 @@
-
-
-class SummaryModel:
-    def compute_loss(self, parameters=None, observables=None):
-        raise NotImplementedError
-
